Adding TableField and SelectionField using refactored Field and Filterbox
