+++ src/components/field.pcss
@@ -1,37 +1,4 @@
 .field {
-  background: #fff;
-  border: 2px solid #d9d9d9;
-  border-radius: 0.4em;
-  box-sizing: border-box;
-  cursor: pointer;
-  display: flex;
-  flex-direction: column;
-  font-size: 1rem;
-  height: 100%;
-  min-width: 22em;
-  padding: 1em;
-  position: relative;
-  z-index: 1;
-
-  &:hover {
-    box-shadow: 0 0 2em -0.5em #999;
-  }
-
-  &.NOT_KEY {
-    background-color: #fff;
-  }
-
-  &.all-excluded {
-    background-color: #e8e8e8;
-  }
-
-  &.single-hit {
-    box-shadow: inset 0 0 4px #ff0;
-  }
-
-  &.filtered {
-    box-shadow: inset 0 0 2px #0f0;
-  }
 
   .name,
   .gwg {
@@ -43,7 +10,7 @@
   }
 
   .name {
-    display: inline-block;
+    /* display: inline-block; */
     font-size: 1.5em;
     font-weight: 600;
     line-height: 2rem;
@@ -70,21 +37,5 @@
   .bartext {
     font-size: 1.375em;
     line-height: 2rem;
-  }
-
-  .details {
-    box-sizing: border-box;
-    display: flex;
-    flex: 1 0 auto;
-    position: relative;
-
-    .filterbox {
-      bottom: 0;
-      display: block;
-      left: 0;
-      position: absolute;
-      right: 0;
-      top: 0.5em;
-    }
   }
 }
