+++ src/components/filterbox.pcss
@@ -6,6 +6,7 @@
   outline: none;
   position: relative;
   user-select: none;
+  height: 100%;
 
   .item {
     cursor: pointer;
