+++ src/components/model.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import usePromise from 'react-use-promise';
 
 import ScrollArea from './scroll-area';
-import Field from './field';
+import TableField from './table-field';
 import logic from '../logic/logic';
 import atplay from '../logic/atplay';
 
@@ -145,7 +145,7 @@ export default function Model({ app, appLayout }) {
                   key={`${tableName}:${fieldName}`}
                   fieldz={fieldName}
                 >
-                  <Field app={app} field={fieldName} fieldData={x} showFilterbox={isFilterboxOpen} />
+                  <TableField app={app} field={fieldName} fieldData={x} showFilterbox={isFilterboxOpen} />
                   {x.subsetRatioText ? (
                     <div className="subsetratio" title={x.subsetRatioTitle}>{x.subsetRatioText}</div>
                   ) : null}
@@ -223,7 +223,7 @@ export default function Model({ app, appLayout }) {
                   key={field.qName}
                   style={cellContainerStyle}
                 >
-                  <Field app={app} field={field.qName} fieldData={field} showFilterbox={isFilterboxOpen} />
+                  <TableField app={app} field={field.qName} fieldData={field} showFilterbox={isFilterboxOpen} />
                 </div>
               );
             }
