+++ src/components/selection-field.pcss
@@ -1,0 +1,124 @@
+.selectionfield {
+  border: none;
+  margin-left: 1em;
+  box-shadow: 0 1px 2px 0 rgba(0,0,0,0.15);
+  height: 5em;
+  padding: .2em;
+  display: flex;
+  position: relative;
+  z-index: 1;
+  flex-direction: column;
+  min-width: 22em;
+  font-size: 1rem;
+  background: #fff;
+  cursor: pointer;
+  border-radius: 0.4em;
+  box-sizing: border-box;
+}
+
+.content {
+  position: absolute;
+  background-color: #f9f9f9;
+  min-width: 160px;
+  box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
+  z-index: 4;
+}
+
+.content {	z-index: 7;
+  position: absolute; height: 156px;	width: 212px;	border-radius: 2px;	background-color: #FFFFFF;	box-shadow: 0 0 0 1px rgba(0,0,0,0.05), 0 4px 10px 0 rgba(0,0,0,0.15);}
+.dropdown {
+  float: left;
+  overflow: hidden;
+}
+
+.content:before {
+  position: absolute;
+  z-index: -1;
+  content: '';
+  right: calc(50% - 10px);
+  top: -8px;
+  border-style: solid;
+  border-width: 0 10px 10px 10px;
+  border-color: transparent transparent #BFBFBF transparent;
+  transition-duration: 0.3s;
+  transition-property: transform;
+}
+.content:after {
+	bottom: 100%;
+	left: 50%;
+	border: solid transparent;
+	content: " ";
+	height: 0;
+	width: 0;
+	position: absolute;
+	pointer-events: none;
+	border-color: rgba(136, 183, 213, 0);
+	border-bottom-color: #88b7d5;
+	border-width: 30px;
+	margin-left: -30px;
+}
+
+.arrow_box {
+position: relative;
+background: #88b7d5;
+}
+
+.arrow_box {
+padding: 40px;
+width: 280px;
+height: 100px;
+border-radius: 6px;
+}
+
+.arrow_box:after {
+	bottom: 100%;
+	left: 50%;
+	border: solid transparent;
+	content: " ";
+	height: 0;
+	width: 0;
+	position: absolute;
+	pointer-events: none;
+	border-color: rgba(136, 183, 213, 0);
+	border-bottom-color: #88b7d5;
+	border-width: 10px;
+	margin-left: -10px;
+}
+
+.popover-wrapper {
+  position: relative;
+  display: inline-block;
+}
+.popover-content {
+  opacity: 0;
+  visibility: hidden;
+  position: absolute;
+  transform: translate(0,10px);
+  background-color: #BFBFBF;
+  padding: 1.5rem;
+  box-shadow: 0 2px 5px 0 rgba(0, 0, 0, 0.26);
+  width: auto;
+}
+.popover-content:before {
+  position: absolute;
+  z-index: -1;
+  content: '';
+  right: calc(50% - 10px);
+  top: -8px;
+  border-style: solid;
+  border-width: 0 10px 10px 10px;
+  border-color: transparent transparent #FFFFFF transparent;
+  transition-duration: 0.3s;
+  transition-property: transform;
+}
+.popover-content {
+  width: 100%;
+  height: 12em;
+  z-index: 10;
+  opacity: 1;
+  visibility: visible;
+  transition: all 0.5s cubic-bezier(0.75, -0.02, 0.2, 0.97);
+  border-radius: 2px;	
+  background-color: #FFFFFF;
+  box-shadow: 0 0 0 1px rgba(0,0,0,0.05), 0 4px 10px 0 rgba(0,0,0,0.15);
+}
