+++ src/components/selections.pcss
@@ -26,12 +26,4 @@
     height: 100%;
     padding: 1em;
   }
-
-  .field {
-    border: none;
-    box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.15);
-    height: 5em;
-    margin-left: 1em;
-    padding: 0.2em;
-  }
 }
