+++ src/components/table-field.pcss
@@ -1,0 +1,53 @@
+.tablefield {
+  display: flex;
+  position: relative;
+  z-index: 1;
+  flex-direction: column;
+  min-width: 22em;
+  height: 100%;
+  font-size: 1rem;
+  background: #fff;
+  cursor: pointer;
+  border-radius: 0.4em;
+  box-sizing: border-box;
+  padding: 1em;
+  border: 2px solid #d9d9d9;
+
+  &:hover {
+    box-shadow: 0 0 2em -0.5em #999;
+  }
+
+  &.NOT_KEY {
+    background-color: #fff;
+  }
+
+  &.all-excluded {
+    background-color: #e8e8e8;
+  }
+
+  &.single-hit {
+    box-shadow: inset 0 0 4px #ff0;
+  }
+
+  &.filtered {
+    box-shadow: inset 0 0 2px #0f0;
+  }
+  
+
+.details {
+  box-sizing: border-box;
+  display: flex;
+  flex: 1 0 auto;
+  position: relative;
+
+  .filterbox {
+    bottom: 0;
+    display: block;
+    left: 0;
+    position: absolute;
+    right: 0;
+    top: 0.5em;
+  }
+}
+
+}
