+++ src/components/topbar.pcss
@@ -7,7 +7,6 @@
   flex-shrink: 0;
   font-size: 1rem;
   height: 6em;
-  overflow: auto;
   position: relative;
   width: 100%;
